content-security-policy/securitypolicyviolation/source-file-blob-scheme.html

<!DOCTYPE html>
<meta http-equiv="Content-Security-Policy" content="script-src 'nonce-abc' blob:">
<script nonce="abc" src="/resources/testharness.js"></script>
<script nonce="abc" src="/resources/testharnessreport.js"></script>
<script nonce="abc">

async_test(t => {
  const scriptText = `
    try {
      eval("assert_unreached('no eval')");
    } catch (e) {
      assert_equals(e.name, 'EvalError');
    }
  `;

  const script = document.createElement("script");
  script.src = URL.createObjectURL(new Blob([scriptText], {type: "text/javascript"}));

  const watcher = new EventWatcher(t, document, 'securitypolicyviolation');
  watcher.wait_for('securitypolicyviolation').then(t.step_func_done(e => {
    assert_equals(e.blockedURI, "eval");
    assert_equals(e.sourceFile, script.src);
    // [3:7] is using a 1-based line and column numbers. The reference the 'e'
    // from the 'eval' word in `scriptText`.
    assert_equals(e.lineNumber, 3);
    assert_equals(e.columnNumber, 7);
  }));

  document.head.append(script);
}, "Violations from blob:-URL scripts have a sourceFile matching the full URL");

</script>
